Add support for multiple bundlers and module systems (#3)
To add support for multiple module systems, the project needs to be compiled in both Commonjs and ESM formats. tsup will now be used for all compilation purposes. The setup for compilation during development and production is different to minimize workflow disruptions. In development, we only compile to ESM while for the package distribution, we compile to both module systems together using tsup's multi-format transpilation.

This project is now an ESM module, however, we still support commonjs projects and have been tested to work with Nextjs build tools. One can both require or import the project's package without issues.

Multi-bundler support will now be possible with unplugin. For starters, we only support webpack, but in the future support for other bundlers will added.